requirements.txt
numpy==2.2.3
pandas==2.2.3
plotly==6.0.0
Pyrebase4==4.8.0
scikit_learn==1.6.1
scipy==1.15.2
statsmodels==0.14.4
streamlit==1.42.1
